upgrade to tokio_uds v0.2
It's nicer because it has an API closer to tokio_tcp.

Also fixed a bug (?) in FastcgiTransport where it would panic when
poll_complete() was called on a dead connection. Using the `cgi-fcgi`
tester, I find that this happens a lot, and reading the docs more
closely, I think the code should tolerate it instead of panicking.